Prepare 1.6.31 preflight
