+++ src/player/ipc.rs
@@ -7,10 +7,7 @@
 
 use std::collections::{HashMap, HashSet, VecDeque};
 use std::io;
-use std::sync::{
-    Arc,
-    atomic::{AtomicBool, Ordering},
-};
+use std::sync::{Arc, atomic::AtomicBool};
 use std::time::Instant;
 
 use interprocess::local_socket::GenericFilePath;
@@ -27,10 +24,13 @@ use super::{
 };
 use crate::player::long_form_seek::{CacheAction, CacheEffectiveState};
 
+mod actor_exit;
 mod audio_output;
 mod wire;
 
 pub(super) use wire::write_json;
+
+use actor_exit::{ActorExit, finish_actor, transport_exit_or_shutdown};
 
 #[cfg(test)]
 use audio_output::{
@@ -43,130 +43,6 @@ use audio_output::{
 /// Upper bound on a single mpv IPC line. mpv's JSON events are well under a kilobyte; the
 /// cap only guards against a broken/hostile endpoint growing one line without limit.
 pub(crate) const MPV_IPC_MAX_LINE: usize = 1024 * 1024;
-
-enum ActorExit {
-    CommandChannelClosed,
-    Eof,
-    Read(io::Error),
-    OversizedLine,
-    Write {
-        operation: &'static str,
-        error: io::Error,
-    },
-    SeekCausalityLost,
-    InternalCommandFailed {
-        operation: &'static str,
-        rejected: bool,
-    },
-    CacheEmergency {
-        file_generation: u64,
-        position_secs: f64,
-        paused: bool,
-        reason: crate::player::long_form_seek::CacheReason,
-    },
-}
-
-impl ActorExit {
-    fn barrier_reason(&self) -> String {
-        match self {
-            Self::CommandChannelClosed => {
-                "mpv command channel closed before acknowledgement".to_owned()
-            }
-            Self::Eof => "mpv IPC closed before command acknowledgement".to_owned(),
-            Self::Read(error) => format!("mpv IPC read failed before acknowledgement: {error}"),
-            Self::OversizedLine => {
-                "mpv IPC protocol failed before command acknowledgement".to_owned()
-            }
-            Self::Write { operation, error } => {
-                format!("mpv IPC {operation} write failed before acknowledgement: {error}")
-            }
-            Self::SeekCausalityLost => {
-                "mpv seek completion became ambiguous before acknowledgement".to_owned()
-            }
-            Self::InternalCommandFailed { operation, .. } => {
-                format!("mpv {operation} failed before required command acknowledgement")
-            }
-            Self::CacheEmergency { reason, .. } => format!(
-                "managed cache safety recycle ({}) before acknowledgement",
-                reason.id()
-            ),
-        }
-    }
-
-    fn transport_reason(self) -> Option<String> {
-        let reason = match self {
-            Self::CommandChannelClosed => return None,
-            Self::Eof => "mpv IPC closed unexpectedly".to_owned(),
-            Self::Read(error) => format!("mpv IPC read failed: {error}"),
-            Self::OversizedLine => {
-                format!("mpv IPC message exceeded the {MPV_IPC_MAX_LINE}-byte safety limit")
-            }
-            Self::Write { operation, error } => {
-                format!("mpv IPC {operation} write failed: {error}")
-            }
-            Self::SeekCausalityLost => {
-                "mpv seek completion timed out; recycling ambiguous transport".to_owned()
-            }
-            Self::InternalCommandFailed {
-                operation,
-                rejected,
-            } => format!(
-                "mpv {operation} {}; recycling player transport",
-                if rejected {
-                    "was rejected"
-                } else {
-                    "timed out"
-                }
-            ),
-            Self::CacheEmergency { .. } => return None,
-        };
-        Some(crate::util::sanitize::sanitize_error_text(reason))
-    }
-}
-
-fn finish_actor(exit: ActorExit, emit: &EventSink) {
-    super::diagnostics::actor_closed();
-    match exit {
-        ActorExit::CacheEmergency {
-            file_generation,
-            position_secs,
-            paused,
-            reason,
-        } => {
-            tracing::error!(
-                file_generation,
-                position_secs,
-                paused,
-                reason = reason.id(),
-                "managed cache safety boundary requires player recycle"
-            );
-            emit(PlayerEvent::CacheEmergency {
-                file_generation,
-                position_secs,
-                paused,
-                reason,
-            });
-        }
-        exit => {
-            if let Some(reason) = exit.transport_reason() {
-                tracing::warn!(%reason, "mpv IPC transport closed");
-                emit(PlayerEvent::TransportClosed(reason));
-            }
-        }
-    }
-}
-
-fn transport_exit_or_shutdown(
-    cmd_rx: &Receiver<PlayerCmd>,
-    intentional_close: &AtomicBool,
-    failure: ActorExit,
-) -> ActorExit {
-    if intentional_close.load(Ordering::Acquire) || cmd_rx.is_closed() {
-        ActorExit::CommandChannelClosed
-    } else {
-        failure
-    }
-}
 
 struct DispatchState {
     last_sent_time_sec: Option<i64>,
