+++ src/terminal_runtime/runner.rs
@@ -22,12 +22,14 @@ use super::startup::StartupTrace;
 
 mod buffered_events;
 mod perf_stats;
+mod recorder_status;
 mod runtime_paths;
 mod teardown;
 #[cfg(test)]
 mod teardown_tests;
 use buffered_events::BufferedWorkerEvents;
 use perf_stats::PerfStats;
+use recorder_status::recorder_capacity_blocked_status;
 use runtime_paths::{TerminalRuntimePaths, resolve as terminal_runtime_paths};
 use teardown::{OwnerIngressDrain, OwnerTeardown, complete_owner_teardown};
 
@@ -1508,31 +1510,6 @@ pub async fn run(
         worker_rx: &mut worker_rx,
     };
     complete_owner_teardown(&mut teardown, owner_error).await
-}
-
-fn recorder_capacity_blocked_status(report: &crate::recorder::job::RecoveryReport) -> String {
-    if report.admission_uncertain {
-        let detail = report
-            .warnings
-            .first()
-            .map(String::as_str)
-            .unwrap_or("recovery inventory could not be verified");
-        if crate::i18n::is_korean() {
-            format!("자동 녹음 일시 중지: 복구 저장 목록을 확인할 수 없음 — {detail}")
-        } else {
-            format!("Automatic recording paused: recovery inventory is uncertain — {detail}")
-        }
-    } else if crate::i18n::is_korean() {
-        format!(
-            "자동 녹음 일시 중지: 저장 대기 {}개 / {}바이트",
-            report.pending, report.pending_bytes
-        )
-    } else {
-        format!(
-            "Automatic recording paused: {} pending / {} bytes",
-            report.pending, report.pending_bytes
-        )
-    }
 }
 
 #[cfg(test)]
